# src/main/resources/static/swagger_api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/products/{id}":{"get":{"tags":["product-controller"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"put":{"tags":["product-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/products":{"post":{"tags":["product-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/helloWorldPlain":{"get":{"tags":["hello-world-contoller"],"operationId":"helloWOrldPlain","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/products/search":{"get":{"tags":["product-controller"],"operationId":"search","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"productType","in":"query","required":true,"schema":{"type":"string"}},{"name":"minSalePrice","in":"query","required":true,"schema":{"type":"number","format":"float"}},{"name":"maxSalePrice","in":"query","required":true,"schema":{"type":"number","format":"float"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"orderBy","in":"query","required":false,"schema":{"type":"string","default":"id"}},{"name":"itensPerPage","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"ASC"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}}}}}},"/api/products/list-all":{"get":{"tags":["product-controller"],"operationId":"listAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}}}}}}},"components":{"schemas":{"ProductDTO":{"required":["code","description","name","preparationTime"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"code":{"type":"string"},"costPrice":{"minimum":0.0,"exclusiveMinimum":false,"type":"number","format":"double"},"salePrice":{"minimum":0.0,"exclusiveMinimum":false,"type":"number","format":"double"},"preparationTime":{"type":"string"},"productType":{"type":"string","enum":["SNACK","DRINK","DESSERT"]},"comments":{"type":"string"},"hasActive":{"type":"boolean"}}}}}}